implemented inheritance in CPE matching
